Added Exit Function

+++ codeGenerator/codeGenerator.go
@@ -148,7 +148,7 @@ func (cg *CodeGenerator) generateFunctionCall(node *parser.Node) {
 	if exists {
 		cg.instructions = append(cg.instructions, VM.Instruction{InstructionType: VM.IT_CallBuiltInFunc, InstructionValue: []byte{builtInFunction}})
 	} else if functionCall.Identifier == "exit" {
-		cg.instructions = append(cg.instructions, VM.Instruction{InstructionType: VM.IT_Halt, InstructionValue: []byte{byte(functionCall.Arguments[0].Value.(*parser.LiteralNode).Value.(int))}})
+		cg.instructions = append(cg.instructions, VM.Instruction{InstructionType: VM.IT_Halt, InstructionValue: []byte{byte(functionCall.Arguments[0].Value.(*parser.LiteralNode).Value.(int64))}})
 	}
 }
